ofctrl-seqno: Do not truncate the last acked value
The requested and acked seqno values are allowed
to be uint64_t, however the values that were added
to the hmap were truncated to uint32_t. This would
lead to loss of information when the value is bigger.

Use uin64_t for the function signatures and for the
hash to prevent truncation.

Fixes: c93c626248c1 ("controller: Implement a generic barrier based on ofctrl cur_cfg sync.")
Reported-at: https://bugzilla.redhat.com/2074019
(cherry picked from commit 292717d64f86b601a139164e70b75ba356cd31be)

+++ controller/ofctrl-seqno.h
@@ -37,7 +37,7 @@ struct ofctrl_ack_seqno {
 struct ofctrl_acked_seqnos *ofctrl_acked_seqnos_get(size_t seqno_type);
 void ofctrl_acked_seqnos_destroy(struct ofctrl_acked_seqnos *seqnos);
 bool ofctrl_acked_seqnos_contains(const struct ofctrl_acked_seqnos *seqnos,
-                                  uint32_t val);
+                                  uint64_t val);
 
 void ofctrl_seqno_init(void);
 size_t ofctrl_seqno_add_type(void);
